// first-vite-ts/src/components/types.ts
export interface CoursePartBase {
    name: string;
    exerciseCount: number;
  }

export interface CoursePartBaseWithDescription extends CoursePartBase {
    description: string;
}

export interface CoursePartBasic extends CoursePartBaseWithDescription {
    kind: "basic"
  }
  
export interface CoursePartGroup extends CoursePartBase {
    groupProjectCount: number;
    kind: "group"
  }
  
export interface CoursePartBackground extends CoursePartBaseWithDescription {
    backgroundMaterial: string;
    kind: "background"
  }
  
export type CoursePart = CoursePartBasic | CoursePartGroup | CoursePartBackground;